+++ src/components/spartan/SModalCard/index.ts
@@ -1,1 +1,3 @@
 export { default as SModalCard } from './SModalCard.vue';
+
+export type { TModalCardEmits, TModalCardProps } from './types';